Add: ui provider page, update: body home page, update: custom navigator bar, update: main.
Create ui provider page to get the information depends on where the user select the menu by this way change the page.
Update body home page with the provider using a final to include the UiProvider type and context.
update custom navigator bar page with the same final used in the body home page and add the ontap to avoid the app to change the page.
Update the main wrapping the material app into a multiprovider.
Install provider 6.0.3