Update:  direcciones page, update: body home page, update: scan button page.
Update direcciones page with the method to save all the scans that have the http type.
update body home page with a class that use the provider but with the listen in false to denied the page to re draw when the information change.
update scan button page to incluide a scan geo type.